API layer reorganization: Phase 1 & 2 complete
- Removed old monolithic dependencies.py file
- Created modular dependency structure:
  - dependencies/auth.py for authentication
  - dependencies/repositories.py for repository injection
  - dependencies/services.py for domain service injection
  - dependencies/use_cases.py for use case injection
- Moved webhook_models.py from models/ to schemas/
- Removed models/ directory from API layer
- Updated all imports from api.models to api.schemas
- Updated imports from src.core to src.infrastructure

API layer now follows proper DDD structure with clean separation.
