+++ src/api/routers/health.py
@@ -16,7 +16,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from src.api.schemas.common import HealthCheckResponse, StatusResponse
 from src.core.cache import cache
 from src.core.config import settings
-from src.core.database import get_db
+from src.infrastructure.database import get_db
 from src.services.storage import storage_service
 
 logger = logging.getLogger(__name__)
